[BEAM-59] Use ServiceLoader to register IOChannelFactories in IOChannelUtils.
